requirements.txt
pyzmq==25.1.2
msgpack-python==0.5.6
numpy==1.26.3
scipy==1.12.0
pyserial==3.5